knn color query works but only if defined

+++ search/src/main/scala/weco/api/search/models/RgbColor.scala
@@ -3,7 +3,7 @@ package weco.api.search.models
 import scala.util.Try
 //import java.awt.Color
 
-case class RgbColor(r: Int, g: Int, b: Int)
+case class RgbColor(r: Float, g: Float, b: Float)
 
 object RgbColor {
   def fromHex(color: String): Try[RgbColor] = Try {
